i have introduced amodularity way of the files hence a splash sleen file that shows for 3 seconds. hence have removed the landing page as well so that it only shows the loading page only
