+++ android/app/build.gradle
@@ -4,13 +4,13 @@ plugins {
     id "dev.flutter.flutter-gradle-plugin"
 }
 
-/*def localProperties = new Properties()
+def localProperties = new Properties()
 def localPropertiesFile = rootProject.file('local.properties')
 if (localPropertiesFile.exists()) {
     localPropertiesFile.withReader('UTF-8') { reader ->
         localProperties.load(reader)
     }
-}*/
+}
 
 def flutterVersionCode = localProperties.getProperty('flutter.versionCode')
 if (flutterVersionCode == null) {
