+++ pubspec.yaml
@@ -35,6 +35,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  flutter_spinkit: ^5.2.0
+  google_fonts: ^6.1.0
 
 dev_dependencies:
   flutter_test:
